pyproject.toml
[tool.poetry]
name = "fivvt-backend"
version = "0.1.0"
description = ""
authors = ["zhikh <zhikh.k@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^5.0.1"
drf-yasg = "^1.21.10"
setuptools = "^76.0.0"
drf-spectacular = "^0.28.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
